Fixing startup logic
Removed "exec(open('kmip/version.py').read())" it is unclear what
this was attempting to accomplish but opening a Python file and
executing its content is analogous to the "import" statement.
This was also attempting to open and execute a file using a path
relative to the folder of invocation, this is a security flaw
even if it succeeds in finding such a file.