Remove AUTHORS.txt
The authors included in AUTHORS.txt hasn't been updated in nearly a
year and is outdated. Rather than tracking the authors explicitly, it
would be better to extract this information from Git's commit
history.

The following command can be used to retrieve a list of the authors
for all the patches contributed to PyKMIP:
    git log | grep ^Author | sort | uniq

Resolves #30