Update the kmip package to support importing the client
This change updates the root-level kmip package, allowing users
to now import the client directly from the kmip package:

from kmip import KmipClient

The KmipClient is an alias for the ProxyKmipClient located in
kmip.pie.client. This will simplify usage and help obfuscate
internal package details that may change in the future.